Sync between gerrit and bitbucket (env. provisioning)

+++ jenkins/jobs/dsl/environment_provisioning.groovy
@@ -105,7 +105,6 @@ docker exec proxy /usr/sbin/nginx -s reload
         environmentVariables {
             propertiesFile('endpoints.txt')
         }
-	systemGroovyCommand(readFileFromWorkspace('cartridge/jenkins/scripts/jenkins_global_envs.groovy'))
     }
     scm {
         git {
@@ -145,8 +144,8 @@ nginx_public_env_conf="${PROJECT_NAME_KEY}-public.conf"
 
 # Remove entry from Nginx
 echo "Deleting main Nginx configuration"
-docker exec -i proxy rm /etc/nginx/sites-enabled/${nginx_main_env_conf}
-docker exec -i proxy rm /etc/nginx/sites-enabled/${nginx_public_env_conf}
+docker exec -i proxy bash -c "if test -f /etc/nginx/sites-enabled/${nginx_main_env_conf}; then rm /etc/nginx/sites-enabled/${nginx_main_env_conf}; fi"
+docker exec -i proxy bash -c "if test -f /etc/nginx/sites-enabled/${nginx_public_env_conf}; then rm /etc/nginx/sites-enabled/${nginx_public_env_conf}; fi"
 
 for node_name in ${node_names_list[@]}; do
     SITE_NAME=$(echo ${node_name} | sed "s/NodeApp//g")
@@ -160,10 +159,14 @@ for node_name in ${node_names_list[@]}; do
 
     echo "Deleting Nginx configuation and removing Docker container for ${node_name}"
     nginx_sites_enabled_file="${PROJECT_NAME_KEY}-$(echo ${node_name} | tr '[:upper:]' '[:lower:]').conf"
-    docker exec -i proxy rm /etc/nginx/sites-enabled/${nginx_sites_enabled_file}
+    docker exec -i proxy bash -c "if test -f /etc/nginx/sites-enabled/${nginx_sites_enabled_file}; then rm /etc/nginx/sites-enabled/${nginx_sites_enabled_file}; fi"
+
     container_id=$(docker ps --format "{{.ID}}: {{.Names}}" | grep ${full_node_name} | cut -f1 -d":")
-    docker stop ${container_id%?}
-    docker rm -f ${container_id%?}
+
+    if [ ! -z "$container_id" ]; then
+        docker stop ${container_id%?}
+        docker rm -f ${container_id%?}
+    fi
 done
 
 # Reload Nginx configuration
